rangefeed: introduce processor interface
This PR extracts externally visible processor functions into interface.
This will allow us to provide alternative implementations while keeping
an option to fall back to old one.

Release note: None
