+++ pkg/kv/kvserver/rangefeed/BUILD.bazel
@@ -6,6 +6,7 @@ go_library(
         "budget.go",
         "catchup_scan.go",
         "filter.go",
+        "legacy_processor.go",
         "metrics.go",
         "processor.go",
         "registry.go",
@@ -48,6 +49,7 @@ go_test(
         "budget_test.go",
         "catchup_scan_bench_test.go",
         "catchup_scan_test.go",
+        "legacy_processor_test.go",
         "processor_test.go",
         "registry_test.go",
         "resolved_timestamp_test.go",
