+++ pkg/kv/kvserver/rangefeed/processor_test.go
@@ -145,7 +145,7 @@ const testProcessorEventCCap = 16
 
 func newTestProcessorWithTxnPusher(
 	t *testing.T, rtsIter storage.SimpleMVCCIterator, txnPusher TxnPusher,
-) (*Processor, *stop.Stopper) {
+) (*LegacyProcessor, *stop.Stopper) {
 	t.Helper()
 	stopper := stop.NewStopper()
 
@@ -162,11 +162,11 @@ func newTestProcessorWithTxnPusher(
 		PushTxnsInterval:     pushTxnInterval,
 		PushTxnsAge:          pushTxnAge,
 		EventChanCap:         testProcessorEventCCap,
-		CheckStreamsInterval: 10 * time.Millisecond,
 		Metrics:              NewMetrics(),
 	})
 	require.NoError(t, p.Start(stopper, makeIntentScannerConstructor(rtsIter)))
-	return p, stopper
+	impl := p.(*LegacyProcessor)
+	return impl, stopper
 }
 
 func makeIntentScannerConstructor(rtsIter storage.SimpleMVCCIterator) IntentScannerConstructor {
@@ -178,7 +178,7 @@ func makeIntentScannerConstructor(rtsIter storage.SimpleMVCCIterator) IntentScan
 
 func newTestProcessor(
 	t *testing.T, rtsIter storage.SimpleMVCCIterator,
-) (*Processor, *stop.Stopper) {
+) (*LegacyProcessor, *stop.Stopper) {
 	t.Helper()
 	return newTestProcessorWithTxnPusher(t, rtsIter, nil /* pusher */)
 }
@@ -434,33 +434,6 @@ func TestProcessorBasic(t *testing.T) {
 	require.False(t, r3OK)
 }
 
-func TestNilProcessor(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-	ctx := context.Background()
-	var p *Processor
-
-	// All of the following should be no-ops.
-	require.Equal(t, 0, p.Len())
-	require.NotPanics(t, func() { p.Stop() })
-	require.NotPanics(t, func() { p.StopWithErr(nil) })
-	require.NotPanics(t, func() { p.ConsumeLogicalOps(ctx) })
-	require.NotPanics(t, func() { p.ConsumeLogicalOps(ctx, make([]enginepb.MVCCLogicalOp, 5)...) })
-	require.NotPanics(t, func() { p.ForwardClosedTS(ctx, hlc.Timestamp{}) })
-	require.NotPanics(t, func() { p.ForwardClosedTS(ctx, hlc.Timestamp{WallTime: 1}) })
-
-	// The following should panic because they are not safe
-	// to call on a nil Processor.
-	stopper := stop.NewStopper()
-	defer stopper.Stop(ctx)
-	require.Panics(t, func() { _ = p.Start(stopper, nil) })
-	require.Panics(t, func() {
-		var done future.ErrorFuture
-		p.Register(roachpb.RSpan{}, hlc.Timestamp{}, nil, false, nil,
-			func() {}, &done,
-		)
-	})
-}
-
 func TestProcessorSlowConsumer(t *testing.T) {
 	defer leaktest.AfterTest(t)()
 	p, stopper := newTestProcessor(t, nil /* rtsIter */)
@@ -565,6 +538,7 @@ func TestProcessorMemoryBudgetExceeded(t *testing.T) {
 
 	stopper := stop.NewStopper()
 	var pushTxnInterval, pushTxnAge time.Duration = 0, 0 // disable
+	m := NewMetrics()
 	p := NewProcessor(Config{
 		AmbientContext:       log.MakeTestingAmbientCtxWithNewTracer(),
 		Clock:                hlc.NewClockForTesting(nil),
@@ -572,8 +546,7 @@ func TestProcessorMemoryBudgetExceeded(t *testing.T) {
 		PushTxnsInterval:     pushTxnInterval,
 		PushTxnsAge:          pushTxnAge,
 		EventChanCap:         testProcessorEventCCap,
-		CheckStreamsInterval: 10 * time.Millisecond,
-		Metrics:              NewMetrics(),
+		Metrics:              m,
 		MemBudget:            fb,
 		EventChanTimeout:     time.Millisecond,
 	})
@@ -593,7 +566,7 @@ func TestProcessorMemoryBudgetExceeded(t *testing.T) {
 		func() {},
 		&r1Done,
 	)
-	p.syncEventAndRegistrations()
+	syncEventAndRegistrations(p)
 
 	// Block it.
 	unblock := r1Stream.BlockSend()
@@ -614,16 +587,16 @@ func TestProcessorMemoryBudgetExceeded(t *testing.T) {
 	}
 
 	// Unblock stream processing part to consume error.
-	p.syncEventAndRegistrations()
+	syncEventAndRegistrations(p)
 
 	// Unblock the 'send' channel. The events should quickly be consumed.
 	unblock()
 	unblock = nil
-	p.syncEventAndRegistrations()
+	syncEventAndRegistrations(p)
 
 	require.Equal(t, newErrBufferCapacityExceeded().GoError(), waitErrorFuture(&r1Done))
-	require.Equal(t, 0, p.reg.Len(), "registration was not removed")
-	require.Equal(t, int64(1), p.Metrics.RangeFeedBudgetExhausted.Count())
+	require.Equal(t, 0, p.Len(), "registration was not removed")
+	require.Equal(t, int64(1), m.RangeFeedBudgetExhausted.Count())
 }
 
 // TestProcessorMemoryBudgetReleased that memory budget is correctly released.
@@ -641,7 +614,6 @@ func TestProcessorMemoryBudgetReleased(t *testing.T) {
 		PushTxnsInterval:     pushTxnInterval,
 		PushTxnsAge:          pushTxnAge,
 		EventChanCap:         testProcessorEventCCap,
-		CheckStreamsInterval: 10 * time.Millisecond,
 		Metrics:              NewMetrics(),
 		MemBudget:            fb,
 		EventChanTimeout:     15 * time.Minute, // Enable timeout to allow consumer to process
@@ -663,7 +635,7 @@ func TestProcessorMemoryBudgetReleased(t *testing.T) {
 		func() {},
 		&r1Done,
 	)
-	p.syncEventAndRegistrations()
+	syncEventAndRegistrations(p)
 
 	// Write entries and check they are consumed so that we could write more
 	// data than total budget if inflight messages are within budget.
@@ -674,7 +646,7 @@ func TestProcessorMemoryBudgetReleased(t *testing.T) {
 			hlc.Timestamp{WallTime: int64(i + 2)},
 			[]byte("value")))
 	}
-	p.syncEventAndRegistrations()
+	syncEventAndRegistrations(p)
 
 	// Count consumed values
 	consumedOps := 0
@@ -683,7 +655,7 @@ func TestProcessorMemoryBudgetReleased(t *testing.T) {
 			consumedOps++
 		}
 	}
-	require.Equal(t, 1, p.reg.Len(), "registration was removed")
+	require.Equal(t, 1, p.Len(), "registration was removed")
 	require.Equal(t, 10, consumedOps)
 }
 
@@ -1072,14 +1044,14 @@ func TestProcessorRegistrationObservesOnlyNewEvents(t *testing.T) {
 // syncEventAndRegistrations waits for all previously sent events to be
 // processed *and* for all registration output loops to fully process their own
 // internal buffers.
-func (p *Processor) syncEventAndRegistrations() {
+func (p *LegacyProcessor) syncEventAndRegistrations() {
 	p.syncEventAndRegistrationSpan(all)
 }
 
 // syncEventAndRegistrations waits for all previously sent events to be
 // processed *and* for all registration output loops for registrations
 // overlapping the given span to fully process their own internal buffers.
-func (p *Processor) syncEventAndRegistrationSpan(span roachpb.Span) {
+func (p *LegacyProcessor) syncEventAndRegistrationSpan(span roachpb.Span) {
 	syncC := make(chan struct{})
 	ev := getPooledEvent(event{sync: &syncEvent{c: syncC, testRegCatchupSpan: &span}})
 	select {
@@ -1131,7 +1103,6 @@ func TestBudgetReleaseOnProcessorStop(t *testing.T) {
 		PushTxnsInterval:     pushTxnInterval,
 		PushTxnsAge:          pushTxnAge,
 		EventChanCap:         channelCapacity,
-		CheckStreamsInterval: 10 * time.Millisecond,
 		MemBudget:            fb,
 		Metrics:              NewMetrics(),
 	})
@@ -1153,7 +1124,7 @@ func TestBudgetReleaseOnProcessorStop(t *testing.T) {
 		&done,
 	)
 	rErrC := notifyWhenDone(&done)
-	p.syncEventAndRegistrations()
+	syncEventAndRegistrations(p)
 
 	for i := 0; i < totalEvents; i++ {
 		p.ConsumeLogicalOps(ctx, writeValueOpWithKV(
@@ -1222,7 +1193,6 @@ func TestBudgetReleaseOnLastStreamError(t *testing.T) {
 		PushTxnsInterval:     pushTxnInterval,
 		PushTxnsAge:          pushTxnAge,
 		EventChanCap:         channelCapacity,
-		CheckStreamsInterval: 10 * time.Millisecond,
 		MemBudget:            fb,
 		Metrics:              NewMetrics(),
 	})
@@ -1244,7 +1214,7 @@ func TestBudgetReleaseOnLastStreamError(t *testing.T) {
 		&done,
 	)
 	rErrC := notifyWhenDone(&done)
-	p.syncEventAndRegistrations()
+	syncEventAndRegistrations(p)
 
 	for i := 0; i < totalEvents; i++ {
 		p.ConsumeLogicalOps(ctx, writeValueOpWithKV(
@@ -1302,7 +1272,6 @@ func TestBudgetReleaseOnOneStreamError(t *testing.T) {
 		PushTxnsInterval:     pushTxnInterval,
 		PushTxnsAge:          pushTxnAge,
 		EventChanCap:         channelCapacity,
-		CheckStreamsInterval: 10 * time.Millisecond,
 		MemBudget:            fb,
 		Metrics:              NewMetrics(),
 	})
@@ -1337,7 +1306,7 @@ func TestBudgetReleaseOnOneStreamError(t *testing.T) {
 		func() {},
 		&r2Done,
 	)
-	p.syncEventAndRegistrations()
+	syncEventAndRegistrations(p)
 
 	for i := 0; i < totalEvents; i++ {
 		p.ConsumeLogicalOps(ctx, writeValueOpWithKV(
@@ -1367,7 +1336,8 @@ func TestBudgetReleaseOnOneStreamError(t *testing.T) {
 // stop and drain remaining allocations.
 // We use account and not a monitor for those checks because monitor doesn't
 // necessary return all data to the pool until processor is stopped.
-func requireBudgetDrainedSoon(t *testing.T, processor *Processor, stream *consumer) {
+func requireBudgetDrainedSoon(t *testing.T, s Processor, stream *consumer) {
+	processor := s.(*LegacyProcessor)
 	testutils.SucceedsSoon(t, func() error {
 		processor.MemBudget.mu.Lock()
 		used := processor.MemBudget.mu.memBudget.Used()
@@ -1474,7 +1444,6 @@ func BenchmarkProcessorWithBudget(b *testing.B) {
 		PushTxnsInterval:     pushTxnInterval,
 		PushTxnsAge:          pushTxnAge,
 		EventChanCap:         benchmarkEvents * b.N,
-		CheckStreamsInterval: 10 * time.Millisecond,
 		Metrics:              NewMetrics(),
 		MemBudget:            budget,
 		EventChanTimeout:     time.Minute,
@@ -1496,7 +1465,7 @@ func BenchmarkProcessorWithBudget(b *testing.B) {
 		func() {},
 		&r1Done,
 	)
-	p.syncEventAndRegistrations()
+	syncEventAndRegistrations(p)
 
 	b.ResetTimer()
 	for bi := 0; bi < b.N; bi++ {
@@ -1508,12 +1477,17 @@ func BenchmarkProcessorWithBudget(b *testing.B) {
 		}
 	}
 
-	p.syncEventAndRegistrations()
+	syncEventAndRegistrations(p)
 
 	// Sanity check that subscription was not dropped.
-	if p.reg.Len() == 0 {
+	if p.Len() == 0 {
 		require.NoError(b, waitErrorFuture(&r1Done))
 	}
+}
+
+func syncEventAndRegistrations(p Processor) {
+	impl := p.(*LegacyProcessor)
+	impl.syncEventAndRegistrations()
 }
 
 // TestSizeOfEvent tests the size of the event struct. It is fine if this struct
